Use delayTrackerStyle by default
