+++ src/common/utils/index.js
@@ -1,4 +1,5 @@
 export { default as getMapboxStyleUrl } from './getMapboxStyleUrl';
 export { default as getMapboxMapCopyrights } from './getMapboxMapCopyrights';
 export { default as removeDuplicate } from './removeDuplicate';
-export { default as trackerStyle } from './trackerStyle';
+export { default as delayTrackerStyle } from './delayTrackerStyle';
+export { default as simpleTrackerStyle } from './simpleTrackerStyle';
